requirements.txt
gradio
gunicorn
psycopg2-binary
uvicorn
google-generativeai
pydantic
fastapi
pandas
plotly
matplotlib
Babel